Firmware 1.72b (small bugfixes)
Small Bugfixes + change for Hardware Rev1.1

+++ Inc/main.h
@@ -156,7 +156,7 @@ uint8_t alarm_month;
 uint8_t alarm_weekday;
 
 uint8_t shutdown_flag;
-uint8_t shutdown_time_counter;
+uint16_t shutdown_time_counter;
 uint8_t alarm_shutdown_time_counter;
 
 uint8_t alarmPoweroff_flag;
